pyproject.toml
[project]
name = "ogtool-demo"
version = "0.1.0"
description = "OGtool demo scraper solution"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "bs4>=0.0.2",
    "fastapi>=0.115.12",
    "mistralai>=1.8.2",
    "pydantic-settings>=2.9.1",
    "python-dotenv>=1.1.0",
    "python-multipart>=0.0.20",
    "tavily-python>=0.7.6",
    "uvicorn>=0.34.3",
]

[dependency-groups]
dev = [
    "ipykernel>=6.29.5",
]
